requirements.txt
jinja2
jsonschema
exceptiongroup
pygments
stack_data
typing-extensions
numpy
pandas
matplotlib
seaborn
ipykernel
oauthlib
requests
pyarrow==11.0.0
fastparquet
holidays
scikit-learn
pymongo
python-dotenv
pyaml